--- frontend/src/lib/xray/forms/protocols/shared/index.ts
@@ -1,1 +1,0 @@
-export { default as HysteriaMasqueradeForm } from './HysteriaMasqueradeForm';
