+++ frontend/src/lib/xray/forms/transport/index.ts
@@ -1,1 +1,2 @@
 export { default as FinalMaskForm } from './FinalMaskForm';
+export { default as HysteriaMasqueradeForm } from './HysteriaMasqueradeForm';
